# teams/src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Teams microservice REST API Docs
  description: TeamUp Teams microservice
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/cards:
    post:
      tags:
        - Cards
      summary: Create Card REST API
      description: Create a new card
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequestDto'
      responses:
        '201':
          description: HTTP Status CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
    get:
      tags:
        - Cards
      summary: Get All Cards REST API
      description: Get all cards with pagination and optional filtering by ownerId or title
      operationId: getAllCards
      parameters:
        - name: ownerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by owner ID
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: Filter by title
        - name: teamId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by team ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Page size
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagingCardResponseDto'
  /api/v1/cards/fetch:
    get:
      tags:
        - Cards
      summary: Get Card REST API
      description: Get a card by ID
      operationId: getCardById
      parameters:
        - name: cardId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /api/v1/cards/fetchByUser:
    get:
      tags:
        - Cards
      summary: Get Cards by user ID REST API
      description: Get a cards by user ID
      operationId: getCardsByUserId
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: User ID
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponseDto'
  /api/v1/cards/{cardId}:
    put:
      tags:
        - Cards
      summary: Update Card REST API
      description: Update an existing card
      operationId: updateCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequestDto'
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
    delete:
      tags:
        - Cards
      summary: Delete Card REST API
      description: Delete a card by ID
      operationId: deleteCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
...
components:
  schemas:
    CardRequestDto:
      type: object
      required:
        - title
        - ownerId
        - teamId
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
          description: Title of the card
          example: "Project X"
        description:
          type: string
          maxLength: 500
          description: Description of the card
          example: "This is a project about X technology"
        posterUrl:
          type: string
          format: uri
          description: URL to the poster image
          example: "https://example.com/image.jpg"
        ownerId:
          type: integer
          format: int64
          description: ID of the card owner
          example: 1
        teamId:
          type: integer
          format: int64
          description: ID of the team owning the card
          example: 1
    CardResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Card identifier
          example: 1
        title:
          type: string
          description: Title of the card
          example: "Project X"
        description:
          type: string
          description: Description of the card
          example: "This is a project about X technology"
        posterUrl:
          type: string
          format: uri
          description: URL to the poster image
          example: "https://example.com/image.jpg"
        ownerId:
          type: integer
          format: int64
          description: ID of the card owner
          example: 1
        teamId:
          type: integer
          format: int64
          description: ID of the team that owns the card
          example: 1
        createdAt:
          type: string
          format: date-time
          description: Date and time when the card was created
          example: "2023-01-01T10:00:00"
        createdBy:
          type: string
          description: User who created the card
          example: "admin"
        updatedAt:
          type: string
          format: date-time
          description: Date and time when the card was last updated
          example: "2023-01-02T11:00:00"
        updatedBy:
          type: string
          description: User who last updated the card
          example: "admin"
    ResponseDto:
      type: object
      properties:
        statusCode:
          type: string
          description: Status code of the response
          example: "200"
        statusMessage:
          type: string
          description: Status message of the response
          example: "Request processed successfully"
...
    PagingDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 100
        totalPages:
          type: integer
          description: Total number of pages
          example: 10
        number:
          type: integer
          description: Current page number (0-indexed)
          example: 0
        size:
          type: integer
          description: Number of elements per page
          example: 10
    PagingCardResponseDto:
      allOf:
        - $ref: '#/components/schemas/PagingDto'
        - type: object
          properties:
            content:
              type: array
              description: List of cards in the current page
              items:
                $ref: '#/components/schemas/CardResponseDto'
...